requirements.txt
sentence-transformers>=2.2.2
chromadb>=0.4.15
requests>=2.31.0
python-dotenv>=1.0.0
typer>=0.9.0
datasets>=2.14.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
rich>=13.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
pandas>=1.5.0
scipy>=1.10.0
kaleido>=0.2.1

# Streamlit Web Interface Dependencies
streamlit>=1.28.0
streamlit-plotly-events>=0.0.6

# GPU Monitoring Dependencies
GPUtil>=1.4.0
nvidia-ml-py3>=7.352.0
pynvml>=11.4.1
psutil>=5.9.0
